extract logic into new TicketCardSidebar component
