+++ assets/src/domain/eventEditor/ui/tickets/ticketsList/cardView/TicketCardSidebar.tsx
@@ -1,0 +1,48 @@
+import React, { useCallback } from 'react';
+import { __ } from '@wordpress/i18n';
+
+import { CalendarDateSwitcher, EditDateRangeButton } from '@appCalendars/dateDisplay';
+import { getStatusTextLabel } from '@sharedEntities/tickets/helpers';
+import { useTicketsListFilterState } from '@edtrServices/filterState';
+import { getPropsAreEqual } from '@appServices/utilities';
+import { useTicketMutator } from '@edtrServices/apollo/mutations';
+import { useTimeZoneTime } from '@appServices/hooks';
+import type { TicketItemProps } from '../types';
+
+const TicketCardSidebar: React.FC<TicketItemProps> = ({ entity: ticket }) => {
+	const { displayStartOrEndDate } = useTicketsListFilterState();
+	const { updateEntity } = useTicketMutator(ticket.id);
+	const { siteTimeToUtc } = useTimeZoneTime();
+
+	const onEditHandler = useCallback(
+		(dates: string[]): void => {
+			const [start, end] = dates;
+			// convert start & end dates to proper UTC "startDate" and "endDate"
+			const startDate = siteTimeToUtc(new Date(start)).toISOString();
+			const endDate = siteTimeToUtc(new Date(end)).toISOString();
+			updateEntity({ startDate, endDate });
+		},
+		[ticket.cacheId, updateEntity]
+	);
+	const statusText = getStatusTextLabel(ticket);
+
+	return ticket ? (
+		<>
+			<CalendarDateSwitcher
+				displayDate={displayStartOrEndDate}
+				endDate={ticket.endDate}
+				startDate={ticket.startDate}
+			/>
+			<EditDateRangeButton
+				endDate={ticket.endDate}
+				header={__('Edit Ticket Sales Start and End Dates')}
+				onEditHandler={onEditHandler}
+				tooltip={__('edit ticket sales start and end dates')}
+				startDate={ticket.startDate}
+			/>
+			<div className={'ee-ticket-status-label'}>{statusText}</div>
+		</>
+	) : null;
+};
+
+export default React.memo(TicketCardSidebar, getPropsAreEqual(['entity', 'cacheId']));
